requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.0.0
starlette>=0.36.0